Rename checkpoint_utils to optimizer_utils
Summary:
`components/checkpoint_utils.py` holds optimizer state-dict plumbing, not
checkpointing code. Every public function takes a `torch.optim.Optimizer`:

  canonical_fqn
  init_optim_state
  get_flat_optim_state_dict
  load_flat_optim_state_dict

Its only imports are `typing` and `torch`, and all four consumers are optimizer
code -- `components/optimizer.py`, `experiments/torchft/optimizer.py`,
`experiments/rl/actors/trainer.py`, and `tests/unit_tests/test_state_dict_keys.py`.
Nothing under `components/` that does checkpointing imports it.

The name is presumably historical: optimizer state is serialized into checkpoints.
It reads as checkpointing code, which invites grouping it with the checkpointer,
where it does not belong.

Also update the stale reference in `CheckpointManager`'s docstring.

Test Plan:
`pytest tests/unit_tests/test_state_dict_keys.py` passes (6 tests) in a venv
matching the CPU CI job.

`python -m compileall torchtitan tests scripts` is clean, and no references to
`checkpoint_utils` remain.
