+++ tests/unit_tests/flex_shard/test_distributed_muon.py
@@ -13,7 +13,7 @@ from torch.testing._internal.distributed._tensor.common_dtensor import (
     DTensorTestBase,
     with_comms,
 )
-from torchtitan.components.checkpoint_utils import (
+from torchtitan.components.optimizer_utils import (
     get_flat_optim_state_dict,
     init_optim_state,
     load_flat_optim_state_dict,
