+++ torchtitan/experiments/rl/actors/trainer.py
@@ -13,7 +13,7 @@ import torch
 import torchstore as ts
 from monarch.actor import Actor, current_rank, endpoint
 from torchtitan.components.checkpoint import CheckpointManager
-from torchtitan.components.checkpoint_utils import canonical_fqn
+from torchtitan.components.optimizer_utils import canonical_fqn
 from torchtitan.components.loss import BaseLoss, ChunkedLossWrapper
 from torchtitan.components.lr_scheduler import LRSchedulersContainer
 from torchtitan.components.optimizer import OptimizersContainer
